requirements.txt
gemini
openai
python-dotenv
crewai
crewai-tools
pandas
langchain
langchain-google-genai
openpyxl
